remove generics in cache layer
